+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added `--print-template` for generating a starter `SECURITY.md`
 - Added Markdown report output with `--format markdown`
 
 ## 0.1.0